+++ .env
@@ -5,8 +5,8 @@ PORT = 3000
 
 
 # REACT_APP_HOST_API = "http://192.168.1.19:8080/api/v2"
-REACT_APP_HOST_API = "http://10.137.53.229:8080/api/v2"
-# REACT_APP_HOST_API=https://test.api.choira.io/api/v2
+# REACT_APP_HOST_API = "http://10.137.53.229:8080/api/v2"
+REACT_APP_HOST_API=https://test.api.choira.io/api/v2
